Removing Highest Frequency

+++ src/kernel.cu
@@ -7,7 +7,7 @@ using namespace cv;
 using namespace std;
 
 //DCT matrix T obtained from matlab dctmtx(8)
-__constant__ float dctMatrix[BLOCK_SIZE * BLOCK_SIZE] = {
+__constant__ double dctMatrix[BLOCK_SIZE * BLOCK_SIZE] = {
     0.3536, 0.3536, 0.3536, 0.3536, 0.3536, 0.3536, 0.3536, 0.3536,
     0.4904, 0.4157, 0.2778, 0.0975, -0.0975, -0.2778, -0.4157, -0.4904,
     0.4619, 0.1913, -0.1913, -0.4619, -0.4619, -0.1913, 0.1913, 0.4619,
@@ -19,7 +19,7 @@ __constant__ float dctMatrix[BLOCK_SIZE * BLOCK_SIZE] = {
 };
 
 //transposed DCT matrix T' obtained from matlab dctmtx(8) with a transpose
-__constant__ float IdctMatrix[BLOCK_SIZE * BLOCK_SIZE] = {
+__constant__ double IdctMatrix[BLOCK_SIZE * BLOCK_SIZE] = {
     0.3536, 0.4904, 0.4619, 0.4157, 0.3536, 0.2778, 0.1913, 0.0975,
     0.3536, 0.4157, 0.1913, -0.0975, -0.3536, -0.4904, -0.4619, -0.2778,
     0.3536, 0.2778, -0.1913, -0.4904, -0.3536, 0.0975, 0.4619, 0.4157,
@@ -32,13 +32,13 @@ __constant__ float IdctMatrix[BLOCK_SIZE * BLOCK_SIZE] = {
 
 //F(p,q) = T * f(x,y) * T'
 
-__global__ void DCT(int numRows, int numCols, float *d_image, float *f_image) {
+__global__ void DCT(int numRows, int numCols, double *d_image, double *f_image) {
 
     __shared__ float cache[BLOCK_SIZE*BLOCK_SIZE];  
     int y = threadIdx.y + (blockDim.y*blockIdx.y);
     int x = threadIdx.x + (blockDim.x*blockIdx.x);    
 
-    float sum = 0.0f;
+    double sum = 0.0f;
 
     if(y < numRows && x < numCols){
         cache[threadIdx.y*BLOCK_SIZE + threadIdx.x] = d_image[y*numCols + x];
@@ -75,7 +75,7 @@ __global__ void DCT(int numRows, int numCols, float *d_image, float *f_image) {
 }
 
 //(T'*A)*T
-__global__ void IDCT(int numRows, int numCols, float *f_image, float *r_image) {
+__global__ void IDCT(int numRows, int numCols, double *f_image, float *r_image) {
 
     __shared__ float cache[BLOCK_SIZE*BLOCK_SIZE];  
     int y = threadIdx.y + (blockDim.y*blockIdx.y);
